refactor(discord): move message dispatcher to its own module for the dead-export gate
