+++ extensions/discord/src/monitor/message-handler.module-test-helpers.ts
@@ -1,13 +1,12 @@
 // Discord helper module supports message handler.module test helpers behavior.
 import type { MockFn } from "openclaw/plugin-sdk/plugin-test-runtime";
 import { vi } from "vitest";
+import { createDiscordMessageDispatcher as createRealDiscordMessageHandler } from "./message-dispatcher.js";
 import type { DiscordMessageRunQueueTestingHooks } from "./message-run-queue.js";
 
 export const preflightDiscordMessageMock: MockFn = vi.fn();
 export const processDiscordMessageMock: MockFn = vi.fn();
 
-const { createDiscordMessageDispatcher: createRealDiscordMessageHandler } =
-  await import("./message-handler.js");
 type DiscordMessageHandlerParams = Parameters<typeof createRealDiscordMessageHandler>[0];
 type DiscordMessageHandlerTestingHooks = NonNullable<DiscordMessageHandlerParams["testing"]>;
 type PreflightDiscordMessageHook = NonNullable<
